backend/requirements.txt
requests>=2.31.0,<3
PyYAML>=6.0.1,<7
psycopg[binary]>=3.1,<4
